% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advise.R, R/get_map.R
\docType{data}
\name{data_dpt}
\alias{data_dpt}
\title{advise}
\format{An object of class \code{data.frame} with 96 rows and 35 columns.}
\usage{
data_dpt

data_dpt
}
\arguments{
\item{index}{the index considered}

\item{eq}{database with equipment per capita}

\item{ind}{database with indexes}

\item{eq}{type of equipment}

\item{data}{the data set we are treating}
}
\value{
the best model to predict the indicator considered and the map that gies with itt

a map of France with level of equipment
}
\description{
advise

get_map
}
\keyword{datasets}
