% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinnyappUI.R
\docType{data}
\name{eq_per_category}
\alias{eq_per_category}
\title{Shinny app UI}
\format{An object of class \code{data.frame} with 33 rows and 3 columns.}
\usage{
eq_per_category
}
\value{
shinyapp UI
}
\description{
Shinny app UI
}
\keyword{datasets}
