% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advise.R
\name{equipment_stepwise}
\alias{equipment_stepwise}
\title{advise}
\usage{
equipment_stepwise(index, eq = data_dpt, ind = indicateurs)
}
\arguments{
\item{index}{the index considered}

\item{eq}{database with equipment per capita}

\item{ind}{database with indexes}
}
\value{
the best model to predict the indicator considered and the map that gies with itt
}
\description{
advise
}
