% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Name of equipement or Indicator}
\usage{
get_info(eq, code_insee, data = data_dpt)
}
\arguments{
\item{eq}{name of equipment or indicator}

\item{code_insee}{number of department}

\item{data}{dataframe, it contains the equipment by default}
}
\value{
dataframe, it contains the equipment by default
}
\description{
Number of the department
}
