% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Creates a map of France with level of equipment}
\usage{
get_map(eq, data = data_dpt)
}
\arguments{
\item{eq}{a type of equipment}

\item{data}{the data set we are treating}
}
\value{
A map of France in which you can decide on the type of equipment you want to plot and which plots the quantity of equipment per capita and per department. The color scale tells you whether there is a high concentration of this equipment per capita (red), or low (yellow).
}
\description{
choose the type of equipment you are interested in
}
