% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{data}
\name{indicateurs}
\alias{indicateurs}
\title{Predict}
\format{An object of class \code{data.frame} with 96 rows and 4 columns.}
\usage{
indicateurs
}
\arguments{
\item{variable}{indicator you want to predict}

\item{pred}{your predictor equipment}
}
\value{
returns prediction of a certain index
}
\description{
Predict
}
\keyword{datasets}
