% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Visualize the quality of a linear regression model based on chosen predictors}
\usage{
predict(variable, pred)
}
\arguments{
\item{variable}{economic or social indicator you want to predict}

\item{pred}{type of equipment used as predictor}
}
\value{
returns prediction of a certain index
}
\description{
This function implements a linear regression to predict the chosen economic indicator with a wide range of equipment as predictors. On the returned map, red departments are those with high residuals in the chosen model. An overall red map mean the predictors are not explanatory enough.
}
