% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinnyappServer.R
\name{shinnyappServer}
\alias{shinnyappServer}
\title{Shinny App Server}
\usage{
shinnyappServer(input, output, session)
}
\arguments{
\item{input}{gives the input}

\item{output}{gives the output}

\item{session}{gives the session}
}
\value{
shiny app Server
}
\description{
Shinny App Server
}
