% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\docType{data}
\name{data_dpt}
\alias{data_dpt}
\title{get_map}
\format{An object of class \code{data.frame} with 96 rows and 35 columns.}
\usage{
data_dpt
}
\arguments{
\item{eq}{type of equipment}

\item{data}{the data set we are treating}
}
\value{
a map of France with level of equipment
}
\description{
get_map
}
\keyword{datasets}
