% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advise.R
\name{equipment_stepwise}
\alias{equipment_stepwise}
\title{Choose the best linear model to predict a chosen economic or social indicator}
\usage{
equipment_stepwise(index)
}
\arguments{
\item{index}{economic or social indicator you are looking to study}
}
\value{
best equipment predictor(s) for this indicator
}
\description{
Choose amongst models with different types of equipments as predictors the model which minimizes the BIC creterion.
}
